Revert "Revert "Remove `include/grpcpp/impl/codegen/async_generic_service.h` (#30450)" (#30951)"
This reverts commit a81b61113c5fe8c577a0e058dba0b74fa79ef153.
