+++ src/compiler/cpp_generator.cc
@@ -137,7 +137,7 @@ std::string GetHeaderIncludes(grpc_generator::File* file,
     }
     static const char* headers_strs[] = {
         "functional",
-        "grpcpp/impl/codegen/async_generic_service.h",
+        "grpcpp/generic/async_generic_service.h",
         "grpcpp/impl/codegen/async_stream.h",
         "grpcpp/impl/codegen/async_unary_call.h",
         "grpcpp/impl/codegen/client_callback.h",
