+++ test/cpp/codegen/compiler_test_golden
@@ -27,7 +27,7 @@
 #include "src/proto/grpc/testing/compiler_test.pb.h"
 
 #include <functional>
-#include <grpcpp/impl/codegen/async_generic_service.h>
+#include <grpcpp/generic/async_generic_service.h>
 #include <grpcpp/impl/codegen/async_stream.h>
 #include <grpcpp/impl/codegen/async_unary_call.h>
 #include <grpcpp/impl/codegen/client_callback.h>
